requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
python-dotenv==1.0.1

# LLM & Agents
langchain>=1.2.0
langchain-core>=1.3.0
langchain-ollama>=1.1.0
langgraph>=1.1.0

# Vector DB
chromadb==0.5.23

# YAML/HCL Parsing
pyyaml==6.0.2
python-hcl2==5.0.0

# Frontend
streamlit==1.41.1

# HTTP
httpx==0.28.1
aiofiles==24.1.0

# Utilities
rich==13.9.4
tabulate==0.9.0
